Reorganize code and modules to clean up location of external depencies. Apply review feedback
